refactored two elements in Timeline to be located outside the grid element.
