+++ src/components/Timeline/Timeline.tsx
@@ -1,13 +1,13 @@
-import React, { useEffect } from 'react'
+import React, { useEffect, useState } from 'react'
 import styled from 'styled-components'
-
-
-import { IntervalSurface, } from '../../internal'
-import { ActivityType, AreaType } from '../../types'
+import { DateTimeFormatter, ZonedDateTime } from '@js-joda/core'
+import '@js-joda/timezone'
+import { ActivityType } from 'types'
 
 interface Props {
-  value: AreaType[],
-  onChange: (newValue: AreaType[]) => void,
+  value: ActivityType[],
+  areaIndex: number,
+  onChange: (time: any) => void,
   onClick: (e: React.MouseEvent) => void,
   activeArea: number
 }
@@ -18,52 +18,370 @@ interface IntervalType {
   gridNumber: number
 }
 
+interface ActivityTimeStampType {
+  title: string,
+  startTime: string,
+  endTime: string,
+  startInteger: number, 
+  endInteger: number,
+  id: string,
+  overflowLane: number,
+  isPlaced: boolean
+}
+
+type ActivityTimeStampsType = ActivityTimeStampType[]
 
 export const Timeline = ({ value, onChange, onClick, activeArea }: Props) => {
-  let activities = value[activeArea].activities
+  let activities: ActivityType[] = value[activeArea]?.activities
 
-  const handleIntervalSurfaceClick = (parsedTime: string) => {
-    let currentData: AreaType[] = value
+  const [columnCount, set_columnCount] = useState<number>(1)
+  const [renderItems, set_renderItems] =  useState<ActivityTimeStampType[]>()
 
-    let newActivity: ActivityType = {
-      title: '',
-      startTime: parsedTime,
-      endTime: parsedTime,
-      id: '10',
-      people: [
-      ],
+  const intervals: IntervalType[] = new Array(112).fill(0).map((item, index) => (
+    {
+      display:
+        index * 15 % 60 === 0
+          ? index * 15 / 60 > 12 && index * 15 / 60 < 24
+            ? `${(index * 15 / 60) - 12}pm`
+            : index * 15 / 60 == 12
+              ? `${index * 15 / 60}pm`
+              : index * 15 / 60 == 24
+                ? '12am'
+                : index * 15 / 60 > 24
+                ? `${(index * 15) / 60 - 24}am`
+                : `${(index * 15) / 60}am`
+          : ''
+      ,
+      value:
+        index * 15 % 60 === 0
+          ? `${index * 15 / 60}:00`
+          : index * 15 % 60 === 15
+            ? `${Math.round(index * 15 / 60)}:15`
+            : index * 15 % 60 === 30
+              ? `${Math.floor(index * 15 / 60)}:30`
+              : index * 15 % 60 === 45
+                ? `${Math.floor(index * 15 / 60)}:45`
+                : ''
+      ,
+      gridNumber: index
     }
-    currentData[activeArea]?.activities.push(newActivity)
-    onChange(currentData)
+  ))
+
+  const currentActivityTimeStamps: ActivityTimeStampsType = value.map((activity) => {
+    let startTime = ZonedDateTime.parse(activity?.startTime)
+    let endTime = ZonedDateTime.parse(activity?.endTime)
+
+    let formattedStartTime = startTime.format(DateTimeFormatter.ofPattern('HHmm'))
+    let formattedEndTime = endTime.format(DateTimeFormatter.ofPattern('HHmm'))
+
+    return {
+      "title": activity.title,
+      "startTime": activity.startTime,
+      "endTime": activity.endTime,
+      "startInteger": parseInt(formattedStartTime),
+      "endInteger": parseInt(formattedEndTime),
+      "id": activity.id,
+      "overflowLane": 1,
+      "isPlaced": false
+    }
+  })
+
+  let activitiesByTimeStamp = currentActivityTimeStamps.sort((a, b) => a.startInteger - b.startInteger)
+
+  const calculateOverflowLanes = (activitiesByTimeStamp: ActivityTimeStampsType) => {
+    let isConflicted: boolean | null = null
+    let isEmptyArray: boolean | null = null
+    let laneRecord: ActivityTimeStampType[][] = [[]]
+
+    activitiesByTimeStamp.forEach((activity, index) => {
+      
+      if (index === 0) {
+        activity.overflowLane = 1
+        laneRecord[0].push(activity)
+        laneRecord.push([])
+        return
+      } 
+
+      for (let i = 0; i < laneRecord.length; i++){
+        let lane = laneRecord[i]
+        let laneIndex = i + 1
+        
+        if (lane.length === 0) {
+          isEmptyArray = true
+        } else if (activity.isPlaced === true) {
+          break
+        } else {
+          isEmptyArray = false
+        }
+
+        for (let i = lane.length - 1; i > -1; i--){
+          let singleLane = lane[i]
+          const isStartTimeConflict = singleLane.startInteger < activity.startInteger && activity.startInteger < singleLane.endInteger
+          const isEndTimeConflict = singleLane.startTime < activity.endTime && activity.endTime < singleLane.endTime
+          const isStartTimeIdentical = singleLane.startTime === activity.startTime
+          
+          if (isStartTimeConflict || isEndTimeConflict || isStartTimeIdentical) {
+            isConflicted = true
+            break
+          }
+
+          else if (activity.isPlaced === true) {
+            break
+          }
+          else {
+            isConflicted = false
+          }
+        }
+
+        if (!isConflicted)
+        {
+          if (!isEmptyArray) {
+            activity.overflowLane = laneIndex
+            activity.isPlaced = true
+            laneRecord[i]?.push(activity)
+          }
+        }
+        
+        else if (isEmptyArray)
+        
+        {
+          let newLane: any = []
+          laneRecord.push(newLane)
+
+          if (!activity.isPlaced) {
+            laneRecord[i]?.push(activity)
+            activity.overflowLane = laneIndex
+          }
+          
+          activity.isPlaced = true
+          isConflicted = false
+          isEmptyArray = false
+        } 
+      }
+    })
+
+    if (laneRecord !== undefined) {
+
+      let filteredRecord = laneRecord.filter(item => {
+        if (item.length === 0) {
+          return false
+        }
+        return true
+      })
+      laneRecord = filteredRecord
+    }
+
+    set_columnCount(laneRecord.length)
+    set_renderItems(activitiesByTimeStamp)
   }
 
+  useEffect(() => {
+    calculateOverflowLanes(activitiesByTimeStamp)
+  }, [value]) 
+
+  const autoScrollFirstActivity = (value: ActivityType[]): string => {
+    if (value !== undefined) {
+      let firstActivityStartTime = value.reduce((prev, curr) => prev.startTime < curr.startTime ? prev : curr).startTime
+  
+      let firstActivityGridPosition = renderRow(firstActivityStartTime)
+      
+      if (firstActivityGridPosition) {
+        return (firstActivityGridPosition - 5).toString()
+      }
+      return '6'
+    }  
+    return '6'
+  }
+
+  const renderRow = (time: string) => {
+    let parsedHour: string = ZonedDateTime.parse(time).format(DateTimeFormatter.ofPattern('HH:mm'))
+
+    if (parsedHour.charAt(0) === '0') {
+      let zerolessParsedHour = parsedHour.substring(1)
+      let gridObject = intervals.filter(interval => interval.value === zerolessParsedHour)
+
+      return gridObject[0]?.gridNumber + 1
+    }
+    const gridObject = intervals.filter(interval => interval.value === parsedHour)
+
+    return gridObject[0]?.gridNumber + 1
+  }
+
+  const handleClick = (interval: IntervalType, value: any) => {
+    
+    if (value !== undefined) {
+      const a = value[0]?.startTime
+      const b = ZonedDateTime.parse(a)
+      const datePrefix = b.format(DateTimeFormatter.ofPattern('yyyy-MM-dd'))
+      const offSet = b.format(DateTimeFormatter.ofPattern('x:00'))
+      const timeZone = b.format(DateTimeFormatter.ofPattern('VV'))
+
+      const parsedTime: string = ZonedDateTime.parse(`${datePrefix}T${interval.value}:00.000${offSet}[${timeZone}]`, DateTimeFormatter.ISO_ZONED_DATE_TIME).toString()
+      
+      onChange(parsedTime)
+    }
+    return
+  }
+
+  useEffect(() => {
+    if (value !== undefined) {
+      let initScrollElement: string = autoScrollFirstActivity(value)
+      document.getElementById(initScrollElement)?.scrollIntoView({
+        behavior: 'smooth'
+      })
+    }
+  }, [JSON.stringify(value)])
+
   return (
-    <S.Timeline>
-      <IntervalSurface
-        value={activities} 
-        areaIndex={activeArea}
-        onChange={handleIntervalSurfaceClick}
-        onClick={onClick}
-      />
-    </S.Timeline>
+    <S.Container>
+      <S.Grid columnCount={columnCount}>
+          {/* {
+            intervals.map((interval, index) => 
+              <S.TimeDisplay
+                key={index}
+                style={{ gridColumnStart: 1, gridRowStart: index + 1}}>
+                  <S.TimeSpan>
+                    {interval.display}
+                  </S.TimeSpan>
+                </S.TimeDisplay>
+              )
+          }   */}
+          {
+            renderItems !== undefined
+              ? renderItems.map((activity, index) => 
+                  <S.Activity
+                    key={index}
+                    onClick={(e) => onClick(e)}
+                    id={activity.id}
+                    style={{
+                      gridColumnStart: activity.overflowLane + 1,
+                      gridColumnEnd: activity.overflowLane + 2,
+                      gridRowStart: renderRow(activity?.startTime),
+                      gridRowEnd: renderRow(activity?.endTime)
+                  }}>
+                    <S.ActivityTitle>
+                      {activity.title}
+                    </S.ActivityTitle>
+                  </S.Activity>
+                )
+              : <></>
+              
+          }
+      </S.Grid>
+
+      <S.TimeStampContainer>
+        {
+          intervals.map((interval, index) => 
+            <S.TimeDisplay
+              key={index}
+              style={{ gridColumnStart: 1, gridRowStart: index + 1}}>
+                <S.TimeSpan>
+                  {interval.display}
+                </S.TimeSpan>
+            </S.TimeDisplay>
+          )
+        }
+      </S.TimeStampContainer>
+         
+      <S.IntervalContainer>
+        {
+        
+        intervals.map((interval, index) =>
+          
+          <S.IntervalBlock
+            key={index}
+            id={index.toString()}
+            onClick={() => handleClick(interval, value)}
+            style={{ gridColumnStart: 2, gridColumnEnd: columnCount + 2, gridRowStart: index === 0 ? 1 : index + 1 }}
+              />
+            )
+        }  
+     
+      </S.IntervalContainer>
+    </S.Container>
   )
 }
 
 const S = {
-  Timeline: styled.div`
-  min-width: 100%;
-  height: calc(100% - 50px);
+  Container: styled.div<{}>`
+    max-width: 100%;
+    overflow-x: auto;
+    background: var(--F_Activity_Backdrop);
+    position: relative;
   `,
-  Header: styled.div<{}>`
-    position: sticky;
+  Grid: styled.div<{
+    columnCount: number
+  }>`
+    position: relative;
+    width: 100%;
+    height: 127rem;
+    display: grid;
+    grid-template-columns: ${props => `3rem repeat(${props.columnCount}, 4rem)`};
+    grid-template-rows: repeat(113, 15px);
+    column-gap: 1px;
+  `,
+  IntervalContainer: styled.div<{}>`
+    position: absolute;
     top: 0;
-    min-width: 100%;
-    width: fit-content;
-    background: var(--F_Background);
+    width: 100%;
+    height: 100%;
+  `,
+  TimeStampContainer: styled.div<{}>`
+    position: absolute;
+    top: 0;
+    width: 3rem;
+    height: 100%;
+    background: var(--F_Activity_Backdrop);
+    z-index: 200;
+  `,
+  TimeDisplay: styled.div<{}>`
+    width: 100%;
+    font-size: var(--F_Font_Size_Label);
+    height: 15px;
     display: flex;
-    justify-content: start;
-    align-items: end;
-    z-index: 1000;
-    padding-top: 0.5rem;
+    justify-content: flex-end;
+
+    :nth-child(1){
+      display: none;
+    }
+  `,
+  TimeSpan: styled.div<{}>`
+    padding-right: 0.5rem;
+    margin-top: 0.4rem;
+  `,  
+  IntervalBlock: styled.div<{}>`
+    box-sizing: border-box;
+    width: 100%;
+    height: 15px;
+    z-index: 1;
+    line-height: 0;
+    :nth-child(1n+1) {
+      
+      border-bottom: 1px solid #e7e7e7;
+    }
+    :nth-Child(2n+2) {
+      border-bottom: 1px solid #d7d7d7;
+    }
+  `,
+  Activity: styled.div<{}>`
+    min-width: 4rem;
+    box-sizing: border-box;
+    position: absolute;
+    width: 100%;
+    height: 100%;
+    background: var(--F_Label_Light_Background_Blue);
+    color: var(--F_Label_Background_Blue);
+    font-size: 12px;
+    padding: 0.15rem;
+    z-index: 100;
+    border-radius: 0.25rem;
+  `,
+  ActivityTitle: styled.div<{}>`
+    width: 100%;
+    text-transform: uppercase; 
+
+    /* writing-mode: vertical-lr;
+    text-orientation: upright;
+    letter-spacing: -1px; */
   `
 }
