+++ src/components/Timeline/index.ts
@@ -1,2 +1,2 @@
+export * from './Timeline2'
 export * from './Timeline'
-export * from './IntervalSurface'
